+++ src/main/java/roomescape/login/service/LoginService.java
@@ -1,0 +1,61 @@
+package roomescape.login.service;
+
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import java.time.Instant;
+import java.util.Date;
+import javax.naming.AuthenticationException;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import roomescape.exceptions.NotFoundException;
+import roomescape.login.dto.LoginRequest;
+import roomescape.member.domain.Email;
+import roomescape.member.domain.Member;
+import roomescape.member.domain.Password;
+import roomescape.member.repository.MemberRepository;
+
+@Service
+public class LoginService {
+
+    private final MemberRepository memberRepository;
+
+    @Value("${security.jwt.token.secret-key}")
+    private String secretKey;
+    @Value("${security.jwt.token.expire-length}")
+    private long validityInMilliseconds;
+
+    public LoginService(MemberRepository memberRepository) {
+        this.memberRepository = memberRepository;
+    }
+    // email, password 보고 존재하는 회원인지 확인, 존재하는 회원이면 토큰 발급해주는 기능
+
+    public String createMemberToken(LoginRequest loginRequest) throws AuthenticationException {
+        Member member = memberRepository.findByEmail(new Email(loginRequest.email()))
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 회원입니다."));
+
+        if (member.isPassword(new Password(loginRequest.password()))) {
+            return parseToToken(member);
+        }
+        throw new AuthenticationException("비밀번호가 일치하지 않습니다.");
+    }
+
+    private String parseToToken(Member member) {
+        Instant issuedAt = Instant.now();
+        Instant expiration = issuedAt.plusSeconds(validityInMilliseconds);
+
+        return Jwts.builder()
+                .setSubject(member.getId().toString())
+                .claim("name", member.getName().name())
+                .claim("email", member.getEmail().email())
+                .claim("role", member.getRole().name())
+                .claim("password", member.getPassword().password())
+                .setIssuedAt(Date.from(issuedAt))
+                .setExpiration(Date.from(expiration))
+                .signWith(Keys.hmacShaKeyFor(secretKey.getBytes()))
+                .compact();
+    }
+
+    // 토큰을 보고 회원 정보로 변환해서 반환해주는 기능 -> MemberRequest로 반환해서 컨트롤러가 이를 받고, MemberService에게 전달해주는 방식?
+
+
+}
