+++ src/main/java/roomescape/member/service/MemberService.java
@@ -4,17 +4,11 @@ import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
-import java.time.Instant;
-import java.util.Date;
 import java.util.List;
 import javax.naming.AuthenticationException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import roomescape.exceptions.NotFoundException;
-import roomescape.login.dto.LoginRequest;
-import roomescape.member.domain.Email;
 import roomescape.member.domain.Member;
-import roomescape.member.domain.Password;
 import roomescape.member.dto.MemberIdNameResponse;
 import roomescape.member.dto.MemberNameResponse;
 import roomescape.member.dto.MemberRequest;
@@ -38,32 +32,6 @@ public class MemberService {
                 .stream()
                 .map(MemberIdNameResponse::new)
                 .toList();
-    }
-
-    public String createMemberToken(LoginRequest loginRequest) throws AuthenticationException {
-        Member member = memberRepository.findByEmail(new Email(loginRequest.email()))
-                .orElseThrow(() -> new NotFoundException("존재하지 않는 회원입니다."));
-
-        if (member.isPassword(new Password(loginRequest.password()))) {
-            return parseToToken(member);
-        }
-        throw new AuthenticationException("비밀번호가 일치하지 않습니다.");
-    }
-
-    private String parseToToken(Member member) {
-        Instant issuedAt = Instant.now();
-        Instant expiration = issuedAt.plusSeconds(validityInMilliseconds);
-
-        return Jwts.builder()
-                .setSubject(member.getId().toString())
-                .claim("name", member.getName().name())
-                .claim("email", member.getEmail().email())
-                .claim("role", member.getRole().name())
-                .claim("password", member.getPassword().password())
-                .setIssuedAt(Date.from(issuedAt))
-                .setExpiration(Date.from(expiration))
-                .signWith(Keys.hmacShaKeyFor(secretKey.getBytes()))
-                .compact();
     }
 
     public MemberNameResponse getMemberNameResponseByToken(String token) throws AuthenticationException {
